move crypt function to own files
